+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 [Full diff](https://github.com/sider/runners/compare/0.46.1...HEAD)
 
 - Support `linter.<id>.dependencies` for Python tools [#2280](https://github.com/sider/runners/pull/2280)
+- **Flake8** Fix `Errno::ENOENT` on multiple configurations [#2281](https://github.com/sider/runners/pull/2281)
 
 ## 0.46.1